+++ packages/calcite-components/src/components/autocomplete-item/autocomplete-item.tsx
@@ -1,3 +1,4 @@
+// @ts-strict-ignore
 import { LitElement, property, createEvent, h, JsxNode } from "@arcgis/lumina";
 import { FlipContext, Scale } from "../interfaces";
 import {
@@ -13,6 +14,7 @@ import {
 } from "../../utils/interactive";
 import { IconNameOrString } from "../icon/interfaces";
 import { guid } from "../../utils/guid";
+import { highlightText } from "../../utils/text";
 import { CSS, SLOTS } from "./resources";
 import { styles } from "./autocomplete-item.scss";
 
@@ -74,6 +76,13 @@ export class AutocompleteItem
   /** Specifies an icon to display at the start of the component. */
   @property({ reflect: true }) iconStart: IconNameOrString;
 
+  /**
+   * Pattern for highlighting text matches.
+   *
+   * @private
+   */
+  @property({ reflect: true }) inputValueMatchPattern: RegExp;
+
   /** Accessible name for the component. */
   @property() label: string;
 
@@ -118,7 +127,8 @@ export class AutocompleteItem
 
   // #region Private Methods
 
-  private handleClick(): void {
+  private handleClick(event: MouseEvent): void {
+    event.preventDefault();
     this.calciteInternalAutocompleteItemSelect.emit();
   }
 
@@ -127,7 +137,7 @@ export class AutocompleteItem
   // #region Rendering
 
   override render(): JsxNode {
-    const { active, description, heading, disabled } = this;
+    const { active, description, heading, disabled, inputValueMatchPattern } = this;
 
     return (
       <InteractiveContainer disabled={disabled}>
@@ -142,8 +152,18 @@ export class AutocompleteItem
           {this.renderIcon("start")}
           <slot name={SLOTS.contentStart} />
           <div class={CSS.contentCenter}>
-            <div class={CSS.heading}>{heading}</div>
-            <div class={CSS.description}>{description}</div>
+            <div class={CSS.heading}>
+              {highlightText({
+                text: heading,
+                pattern: inputValueMatchPattern,
+              })}
+            </div>
+            <div class={CSS.description}>
+              {highlightText({
+                text: description,
+                pattern: inputValueMatchPattern,
+              })}
+            </div>
           </div>
           <slot name={SLOTS.contentEnd} />
           {this.renderIcon("end")}